refactor: game logic util functions
