+++ src/components/utils.ts
@@ -1,5 +1,4 @@
 import {COLUMNS, ROWS} from '../consts';
-import {IAllPositions, IPositionSchema, TDirections} from '../interfaces/api';
 
 export function randomIntFromInterval(min: number, max: number) { // min and max included
     return Math.floor(Math.random() * (max - min) + min);
@@ -13,73 +12,9 @@ export function getRandomColumn() {
     return randomIntFromInterval(0, COLUMNS);
 }
 
-export function getOccupiedByPLayers(positions: IAllPositions, rowOrCol: ('row' | 'col')) {
-    //todo can be a set
-    return Object.values(positions).reduce((prev: (number)[], currentPositions) => {
-        const currentRows = currentPositions.position.map((position: IPositionSchema) => position[rowOrCol]);
-        return [...prev, ...currentRows];
-    }, []);
-}
-
-export function getOccupiedByFood(foods: IPositionSchema[], rowOrCol: ('row' | 'col')) {
-    // todo can be a set
-    return foods.map((food) => food[rowOrCol]);
-}
-
-export function getUnoccupiedPosition(positions: IAllPositions, foods: IPositionSchema[]) {
-    // check that new random position is not clashing with existing players or food
-    let randomRow = getRandomRow();
-    const occupiedPlayerRows = getOccupiedByPLayers(positions, 'row');
-    const occupiedFoodRows = getOccupiedByFood(foods, 'row');
-    const allOccupiedRows = [...occupiedPlayerRows, ...occupiedFoodRows];
-    while (allOccupiedRows.includes(randomRow)) {
-        randomRow = getRandomRow();
-    }
-
-    let randomCol = getRandomColumn();
-    const occupiedPlayerCols = getOccupiedByPLayers(positions, 'col');
-    const occupiedFoodCols = getOccupiedByFood(foods, 'col');
-    const allOccupiedCols = [...occupiedPlayerCols, ...occupiedFoodCols];
-    while (allOccupiedCols.includes(randomCol)) {
-        randomCol = getRandomColumn();
-    }
-
-    return {
-        row: randomRow,
-        col: randomCol
-    };
-}
-
-export function getUpdatedFood(foods: IPositionSchema[], eatenFood: IPositionSchema[]): IPositionSchema[] {
-    const updatedFood: IPositionSchema[] = [];
-
-    for (const food of foods) {
-        for (const eaten of eatenFood) {
-            if (food.row !== eaten.row && food.col !== eaten.row) {
-                updatedFood.push(food);
-            }
-        }
-    }
-
-    return updatedFood;
-}
-
 export function bothArraysEqual(prevColors: string[], allColors: string[]) {
     //https://stackoverflow.com/questions/3115982/how-to-check-if-two-arrays-are-equal-with-javascript
     return prevColors.every(item => allColors.includes(item)) && allColors.every(item => prevColors.includes(item));
-}
-
-export function updateDirection(previousDirection: TDirections, newDirection: TDirections) {
-
-    let updatedDirection = newDirection; // new direction by default
-
-    // snake can only go to three directions, i.e. if going UP cannot go DOWN
-    if (previousDirection === 'UP' && updatedDirection === 'DOWN') updatedDirection = 'UP';
-    if (previousDirection === 'RIGHT' && updatedDirection === 'LEFT') updatedDirection = 'RIGHT';
-    if (previousDirection === 'DOWN' && updatedDirection === 'UP') updatedDirection = 'DOWN';
-    if (previousDirection === 'LEFT' && updatedDirection === 'RIGHT') updatedDirection = 'LEFT';
-
-    return updatedDirection;
 }
